Collision with jQuery width and height methods
As i am extending $.fn. with viewportH and viewportW methods, setting this.width and this.height into those methods does override jquery width and height methods ... too bad !
Using viewportWidth and viewportHeight as variable names does fix the conflict.

We may instantiate the viewport class out of the responsiveimage constructor, and use a setOptions method inside of the constructor,  but your way of doing it with if(viewport) seem mutch cleaner and simplier. so +1 for your solution.