[FEATURE] add frame without vertical margins
This feature adds a new item "No vertical margin" to the Frame in Content Element Layout. Since frames now have a top and bottom margin by default, this is useful to allow a frame to attach to another object if this also has no margin (e.g.: a carousel can be attached to the header with no space between them).