[FEATURE] Introduce creditText constant splitted from copyrightText
Usually a website must display the copyrights owned by the company that also owns the website and could optionally display the credits that are basically the creator/developer but can include a link to a page with additional credits.
This patch makes it easier to use these two texts (copyright and credits) independently inside the templates.