[!!!][FEATURE] Add support for spacer in menu processor - fixes #335
Spacer are now enabled by default for main and sub navigation,
since TYPO3 does allow links to pages from type spacer existing
templates that override the default templates need to be adjusted.

If the usage of spacer is not intended or you do not want do support it,
you can simply disable it in the processor configuration.

page.10.dataProcessing.10.includeSpacer = 0
page.10.dataProcessing.20.includeSpacer = 0