[!!!][FEATURE] Add support for google webfonts - fixes #115
Share font was removed in favor of google webfonts. A new
typoscript less constant "theme-google-webfont" was introduced
that contains the name of the webfont. This constant will
be used as the default font-family and also used to include the
css file include from google.

page.includeCSSLibs {
  googlewebfont = https://fonts.googleapis.com/css?family={$plugin.bootstrap_package.settings.less.theme-google-webfont}:300,400,700
}

To avoid syntax errors we provide a set of webfonts from google
that can be choosen from the constant editor in the basic
settings for the bootstrap package.

- Lora
- Merriweather
- Raleway
- Roboto
- Source Sans Pro
- Ubuntu

NOTE:
You can disable the automatic inclusion of the google webfonts
by unsetting the css file inclusion in your typoscript setup.

page.includeCSSLibs.googlewebfont >