[TASK] use includeJS instead of includeJSLibs
RsaEncryptionEncoder adds its JS files with addJsFile() and bootstrap_package adds its JS files with includeJSLibs. This causes to have 2 different merged JS files in the HTML code so 2 different request.
To have a single request we change the inclusion method to includeJS since includeJSLibs is not needed in this case.