[!!!][TASK] Remove obsolete pagetype popup - fixes #476
The current implementation was broken and non functional, after
reconsidering the support for popups it became clear that the needs
for popups vary a lot and can not be considered as basic feature that
needs to be supported. With this change we are dropping the support
and are not planning to implement this again in the future.