[FEATURE] Introduce new parser for scss css files
Bootstrap Package now also supports parsing of scss files, to keep settings as
simple as possible the option to disable css parsing has changed and is now more
generic. A compatibility layer ensures that the option is still disabled if you
disabled it before, but please head to the extension configuration and disable
the new disableCssProcessing option.