[TASK] Remove install notification for unsupported webservers - fixes #562
TYPO3 v9 has adapted to automaticly provide webserver configuration files on
installation. We keep this functionality as long as we support v8.7 but we remove
the note about webservers were users manually need to ensure the configuration
is correct. Since routing is now part or v9.5 LTS we also stop to suggest the usage
of realurl.