[FEATURE] Add support for highres images (#678)
By popular request, we are now adding high-resolution
options to the existing responsive image rendering. The 
Bootstrap Package does not activate them by default, 
but it can be added relatively easily by configuration.

Example configuration to enable high-resolution support 
for the default variant for devices with a minimum device 
pixel ration from 1.5 and 2.

lib.contentElement.settings.responsiveimages.variants {
    default {
        sizes {
            1\.5x {
                multiplier = 1.5
            }
            2x {
                multiplier = 2
            }
        }
    }
}

The data for image rendering is now extended with a 
sub-set for high-resolution variants. If you did not 
overwrite the default templates, it will just work as 
soon as you add the configuration. If you have 
overwritten the default templates, it will still work as 
before but you need to add the new support for 
high-resolution images yourself. 

Please check the new updated templates and adjust
your code if you want this support.

There is now a new key called sizes in every variant 
that should be rendered. This dataset consists of 
several entries each with a default key for a normal 
rendering (1x). This key will be added automatically 
and is always available. It is not possible to store sizes 
smaller than 1. Each size contains a multiplier, which 
is calculated with the base size.

Example:
Base: 1000px

Minimum Device Pixel Ratio = 1.5
Multiplier = 1.25
Result: 1250px

Minimum Device Pixel Ratio = 2
Multiplier = 1.5
Result: 1500px