[FEATURE] Allow processing of configured file types
To deliver the best results possible, the bootstrap package
has a strict type checking before passing assets to the
correct renderings. Even if your system was capable of
rendering images from PDFs it was not allowed by default
through those type checkings, since a PDF is in the
mimetype group application. That means we were not able
to simply allow these application group since that would
also mean that the system would try to generate a image
from a zip file, like in fluid styled content.

Since there is still a huge demand for PDF previews, we
made the decision to extend the current typechecks. TYPO3
allows you to configure imagefile_ext and mediafile_ext in
its configuration. We are now also allowing the processing
of files with extensions configured in the according lists.

Fixes: #647
Fixes: #576