BUGFIX: Fix icon size after constant change
If for example the constant plugin.bootstrap_package_contentelements.texticon.icon.medium.height  and plugin.bootstrap_package_contentelements.texticon.icon.medium.width is changed to 60 the icon doesnt appear bigger in the frontend even if the img size is 60x60. It stays small. Either remove height and width totally or set it to auto.