[!!!][FEATURE] Introduce embedded frames and improve rendering
The new embedded layout for frames enables more customization. 
Instead of spreading the color block/background options just across 
the whole viewport width we now also reduce it to be contained within 
the content width of the page. In addition, the CSS was optimized to 
utilize CSS-Variables to make adjustments easier. 
Without the need to recompile the CSS.

The text color for frames is not calculated anymore, instead, 
we are giving the user the option to define the color.

SCSS Example:
```
$frame-backgrounds: (
  red: (
    background: #ff0000,
    color: #ffffff,
    link-color: #ffeeee,
    link-hover-color: #ffffff
  )
);
```

CSS Variables Example:
```
.frame-background-dark {
  --frame-color: #ff0000;
  --frame-background: #ffffff;
  --frame-link-color: #ffeeee;
  --frame-link-hover-color: #ffffff;
}
```

Important:
The `layout` field for `tt_content` that was passed to the frame before 
is limited to store integers. To allow more flexible and specific layouts 
we are introducing a new layout field `frame_layout` to allow longer, 
more specific layout options and avoid mapping in general. If you 
utilized the `layout` field from `tt_content` before you will need to 
adjust your CSS and extensions.