[TASK] Provide frontend build custom commands (#1068)
This patch introduces two custom commands for DDEV to simplify the frontend related processes.

## `ddev npm`

Runs npm in the `Build` folder in the web container. Parameters are passed to npm. See also `ddev npm -h`.

## `ddev build`

Runs a script in the `Build` folder in the web container. Before the script is executed a `npm install` is performed which can be skipped by adding the `--no-install` option. Default script is `build` or provide the script name as parameter. See also `ddev build -h`.