[FEATURE] Introduce frame options
To provide more flexibility we introduce
frame options to replace most of the existing
options of frame classes. In contrast to the
hard selection of one option like ruler-before
or ruler-after, the frame options allow to
combine those.

The option indent was dropped completly
and is now a combination of indent-left
and indent-right.

Adding new options

```
TCEFORM {
    tt_content {
        frame_options {
            addItems {
                coolstyle = New cool style
            }
        }
    }
}
```

All options will be prefixed with frame-option-* in the output.