[BUGFIX] Render "hidden" fields as `checkboxToggle` (#1247)
"Item" records provided by EXT:bootstrap_package (e.g. for accordion,
carousel, etc.) are usually rendered in IRRE contexts. There is a wild
bug if an element is disabled ("hidden") that makes it very hard to
enable it again.

The root cause is that checkbox fields are internally
handled as bitmasks, triggering a strange behavior as an enabled
checkbox is rendered as disabled as long the element is disabled.

To overcome this issue until TYPO3 Core's behavior is fixed, "hidden"
fields are now rendered as `checkboxToggle` that don't trigger the bug.