[BUGFIX] Source mapping is now public path agnostic (#1352)
The source maps are no longer inlined in the compiled CSS file.
There‘s a .css.map file for the compiled CSS file instead. Sources
are now included in the map file. This means that sources can now
be moved to non-public folders without losing the possibility
to debug your code.

Related: #1263