[BUGFIX] Handle allowed file extensions for TYPO3 v12 (#1365)
With TYPO3 v12.0, a new TCA-type file was introduced. This type is used for the tt_content.media and tt_content.assets fields. Thus, all column overrides for TYPO3 v12 must match the new TCA config.

This PR solves an issue with columns overrides of media columns concerning allowed file extensions. With the new TCA type, the expected TCA config looks different. In order to solve this issue and still support the previous behavior for TYPO3 v11.5, a new TcaUtility class is introduced. It returns a valid TCA config for allowed and/or disallowed file extensions, depending on the current TYPO3 version.