[TASK] Improve icon selector performance in backend (#1392)
The core select icon wizard blocks the editing form
until all images are loaded. This is a problem when
you display a lot of icons in context of a selector.

This is a workaround, and adapts the current selector
to render all icons inline. It should become a custom
render type at some point.